refactor(shoeprint-report): extract modules, fix WebGL batching and escapeHtml
