--- src-tauri/gen/schemas/capabilities.json
@@ -1,1 +1,0 @@
-{"default":{"identifier":"default","description":"Capability for the main window","local":true,"windows":["main","settings","edit"],"permissions":["core:default","core:path:default","core:event:default","core:window:default","core:window:allow-close","core:window:allow-minimize","core:window:allow-toggle-maximize","core:window:allow-start-dragging","core:app:default","core:image:default","core:resources:default","core:menu:default","core:tray:default",{"identifier":"fs:allow-exists","allow":["**","**/*","/**/*"]},{"identifier":"fs:allow-read-file","allow":["**","**/*","/**/*"]},{"identifier":"fs:allow-read-text-file","allow":["**","**/*","/**/*"]},"fs:allow-write-text-file",{"identifier":"fs:allow-write-file","allow":["**","**/*","/**/*"]},"store:allow-set","store:allow-get","store:allow-save","store:allow-load","store:allow-delete","dialog:allow-open","dialog:allow-save","dialog:allow-message","dialog:allow-ask","dialog:allow-confirm",{"identifier":"shell:allow-execute","allow":[{"args":true,"name":"bin/sourceafis_cli","sidecar":true}]},"store:default","dialog:default","fs:default"]},"desktop-capability":{"identifier":"desktop-capability","description":"","local":true,"permissions":["window-state:default"],"platforms":["macOS","windows","linux"]}}
